requirements.txt
## Flask
## mysql-connector-python

Flask==3.0.3
gunicorn==21.2.0
mysql-connector-python==9.1.0

SQLAlchemy==2.0.34
